Refactor: Rename as_nzs_3000 to as_nzs_3000_tools and move from src to project root for a flat layout
